+++ transformer_engine/pytorch/sequential/utils.py
@@ -205,159 +205,3 @@ class Decorator(Protocol):
 
 def reinterpret_cast(x: Any, t: type[T], /) -> T:
     return x
-
-
-def torch_op(func: Callable[PS, T]) -> Callable[PS, T]:
-    import torch
-    from copy import deepcopy
-    from . import cpp_extensions
-
-    def make_wrapper(func: Callable[..., Any]):
-        def type_name(t: type) -> str:
-            if t.__module__ == "builtins":
-                if isinstance(t, GenericAlias):
-                    return str(t)
-                else:
-                    return t.__name__
-            elif t.__module__ == "transformer_engine.pytorch.sequential.cpp_extensions":
-                return f"cpp_extensions.{t.__name__}"
-            else:
-                return f"{t.__module__}.{t.__name__}"
-
-        def wrap_unwrap_code(arg_name: str, arg_type: type, arg_type_name: str):
-            wrapped_arg_type_name = type_name(wrap_type(arg_type))
-            if arg_type is cpp_extensions.Tensor:
-                w = f"{arg_name}_: {wrapped_arg_type_name} = ({arg_name}.data, {arg_name}.amax, {arg_name}.scale, {arg_name}.scale_inv)\n"
-                u = f"{arg_name}: {arg_type_name} = {arg_type_name}(*{arg_name}_)\n"
-            elif issubclass(arg_type, Enum):
-                w = f"{arg_name}_: {wrapped_arg_type_name} = {arg_name}.value\n"
-                u = f"{arg_name}: {arg_type_name} = {arg_type_name}({arg_name}_)\n"
-            elif arg_type in [int, float, bool, str, torch.Tensor]:
-                w = f"{arg_name}_: {wrapped_arg_type_name} = {arg_name}\n"
-                u = f"{arg_name}: {arg_type_name} = {arg_name}_\n"
-            else:
-                raise NotImplementedError(arg_type_name)
-            return (w, u)
-
-        def wrap_type(arg_type: type):
-            if arg_type is cpp_extensions.Tensor:
-                return tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]
-            elif issubclass(arg_type, Enum):
-                return int
-            elif arg_type in [int, float, bool, str, torch.Tensor]:
-                return arg_type
-            else:
-                raise NotImplementedError(arg_type_name)
-
-        def register_op(func: Callable[..., Any], abstract_impl: Callable[..., Any]):
-            name = f"nvte::{func.__name__}"
-            # Different versions of PyTorch have different ways of registering custom ops
-            try:
-                decl, impl, aimp = (  # type: ignore
-                    torch._custom_ops.custom_op,  # type: ignore
-                    torch._custom_ops.impl,  # type: ignore
-                    torch._custom_ops.impl_abstract,  # type: ignore
-                )
-                decl(name)(func)
-                impl(name)(func)
-                aimp(name)(abstract_impl)
-                return
-            except AttributeError:
-                pass
-            try:
-                decl = torch._custom_op.impl.custom_op  # type: ignore
-                declared = decl(name)(func)  # type: ignore
-                declared.impl("cuda")(func)  # type: ignore
-                declared.impl_abstract()(abstract_impl)  # type: ignore
-                return
-            except AttributeError:
-                pass
-            if not hasattr(register_op, "warned"):  # type: ignore
-                register_op.warned = True  # type: ignore
-                warnings.warn("Unable to find custom_op, decorator has no effect")
-
-        # Dynamically generate code of the wrappers
-        arg_types = get_arg_types(func)
-        arg_names = get_arg_names(func)
-        arg_type_names = list(map(type_name, arg_types))
-        return_type = get_return_type(func)
-        return_type_name = type_name(return_type)
-        outer_sig = f"""({ ','.join(
-            f'{arg_name}: {arg_type_name}'
-            for arg_name, arg_type_name in zip(arg_names, arg_type_names)
-        ) }) -> {return_type_name}"""
-        arg_wrapping_code = ""
-        arg_unwrapping_code = ""
-        for arg_name, arg_type, arg_type_name in zip(
-            arg_names, arg_types, arg_type_names
-        ):
-            w, u = wrap_unwrap_code(arg_name, arg_type, arg_type_name)
-            arg_wrapping_code += w
-            arg_unwrapping_code += u
-        wrapped_args = ",".join(f"{arg_name}_" for arg_name in arg_names)
-
-        result_wrapping_code, result_unwrapping_code = wrap_unwrap_code(
-            "result", return_type, return_type_name
-        )
-
-        wrapped_arg_names = [f"{arg_name}_" for arg_name in arg_names]
-        wrapped_arg_types = [wrap_type(t) for t in arg_types]
-        wrapped_arg_type_names = [type_name(t) for t in wrapped_arg_types]
-        wrapped_return_type = wrap_type(return_type)
-        wrapped_return_type_name = type_name(wrapped_return_type)
-        inner_sig = f"""({ ','.join(
-            f'{arg_name}: {arg_type_name}'
-            for arg_name, arg_type_name in zip(wrapped_arg_names, wrapped_arg_type_names)
-        ) }) -> {wrapped_return_type_name}"""
-        unwrapped_args = ",".join(f"{arg_name}" for arg_name in arg_names)
-
-        source = f"""\
-import torch
-from . import cpp_extensions
-
-def abstract_impl{inner_sig}:
-    {arg_unwrapping_code}
-    func.__globals__["_nvte"] = impostor
-    result: {return_type_name} = func({unwrapped_args})
-    func.__globals__["_nvte"] = cpp_extensions
-    {result_wrapping_code}
-    return result_
-
-def {func.__name__}{inner_sig}:
-    {arg_unwrapping_code}
-    result: {return_type_name} = func({unwrapped_args})
-    {result_wrapping_code}
-    return result_
-
-def outer_wrapper{outer_sig}:
-    {arg_wrapping_code}
-    result_: {wrapped_return_type_name} = {func.__name__}({wrapped_args})
-    {result_unwrapping_code}
-    return result
-"""
-        try:
-            # Swap real cpp_extensions (_nvte) for impostor that does nothing
-            # This is needed so the abstract implementation is traceable by PyTorch Dynamo
-            class NVTEImpostor:
-                def __getattr__(self, attr_name: str) -> Any:
-                    attr = getattr(cpp_extensions, attr_name)
-                    if callable(attr):
-                        return lambda *args, **kwargs: None  # type: ignore
-                    else:
-                        return attr
-
-            # Create op
-            ns = dict(func=func, __name__=__name__, impostor=NVTEImpostor())
-            exec_saving_source(source, ns)
-            op_impl = reinterpret_cast(ns[func.__name__], Callable[..., Any])
-            op_wrap = reinterpret_cast(ns["outer_wrapper"], Callable[PS, T])
-            op_aimp = reinterpret_cast(ns["abstract_impl"], Callable[..., Any])
-            register_op(op_impl, op_aimp)
-
-            return op_wrap
-        except Exception as e:
-            raise RuntimeError(
-                f"Failed to compile wrapper for {func.__name__}. Generated code: \n```\n{source}```"
-            ) from e
-
-    return make_wrapper(func)
